+++ release-notes.md
@@ -7,7 +7,9 @@
 * none
 
 ## Bugs fixed:
-* none
+* [#346](https://github.com/openziti/desktop-edge-win/issues/346) Fixed the UI filtering of services on the Identity detail screen
+* [#348](https://github.com/openziti/desktop-edge-win/issues/348) IP addresses do not need to be added to the NRPT
+* [#349](https://github.com/openziti/desktop-edge-win/issues/349) Too many services can cause the NRPT update to fail
 
 ## Dependency Updates
 * none